+++ js/sdk/src/components/GateCreateModal.js
@@ -33,8 +33,6 @@ const GateCreateModal = ({
     setFile(undefined)
   }
 
-  console.log('metadata :>> ', metadata);
-
   const handleFileUpload = async () => {
     setInProgress(true)
     try {
@@ -203,7 +201,7 @@ const StyledPaper = styled(Paper)(({ theme }) => ({
 const StyledCloseIcon = styled(CloseIcon)(({ theme }) => ({
   position: 'absolute',
   right: theme.spacing(1),
-  top: theme.spacing(2),
+  top: theme.spacing(1),
 }))
 
 export default GateCreateModal
